cheri: wireless: broadcom: Return uintptr_t from dma_getvar
This function returns a pointer cast to an integer and the
caller expects that it is possible to cast this back to a
pointer. Thus us uintptr_t.

+++ drivers/net/wireless/broadcom/brcm80211/brcmsmac/dma.c
@@ -1125,7 +1125,7 @@ void dma_counterreset(struct dma_pub *pub)
 }
 
 /* get the address of the var in order to change later */
-unsigned long dma_getvar(struct dma_pub *pub, const char *name)
+uintptr_t dma_getvar(struct dma_pub *pub, const char *name)
 {
 	struct dma_info *di = container_of(pub, struct dma_info, dma);
